requirements.txt
Flask
pandas
snowflake-connector-python
python-dotenv
groq
matplotlib
